mejoras del servicio

+++ lib/AlbumStorage.php
@@ -79,11 +79,7 @@ class AlbumStorage extends Storage {
             }
             $return['presentations'][] = array('id' => $response['id'], 'name'=>$response['name'], 'presentation_id' => $pres['id'], 'url' => $response['mediaLink']);
             
-            if($pres['name'] == 'cover') {
-                $this->tmpCoverImg = $path;
-            } else {
-                unlink($path);
-            }
+            unlink($path);
         }
         
         return $return;
@@ -96,7 +92,9 @@ class AlbumStorage extends Storage {
         foreach($songs as $song) {
             if(!is_null($song['temporal_file_name'])) {
                 $songPath = $this->config->get('song_tmp_path') . "/" . $song['temporal_file_name'];
-                unlink($songPath);
+                if(file_exists($songPath)) {
+                    unlink($songPath);
+                }
             }
         }
         
@@ -162,103 +160,4 @@ class AlbumStorage extends Storage {
         }
         return $return;
     }
-    
-    /**
-     * 
-     * @param array $album
-     */
-    private function downloadZip($album) {
-        echo "Descargando archivo zip: {$album['static_zip_file_name']}\n";
-        mkdir($this->config->get('zip_tmp_path') . '/' . $album['staticDirectory']);
-        file_put_contents($this->config->get('zip_tmp_path') . '/' .$album['static_zip_file_name'], fopen($album['zipUrl'], 'r'));
-    }
-    
-    private function uploadZip($name, $songs, $folder) {
-        echo "Creando zip\n";
-        
-        $zip = new ZipArchive();
-        $zipName = $name;
-        $filePath = $this->config->get('zip_tmp_path') . '/' .$zipName . '.zip';
-        
-        if ($zip->open($filePath, ZipArchive::CREATE) !== TRUE) {
-            throw new Exception('Error al crear el archivo Zip');
-        }
-
-        foreach($songs as $song) {
-            if(!is_null($song['url'])) {
-                $file = file_get_contents($song['url']);
-                $name = str_replace($folder, '', $song['static_file_name:']);
-                $zip->addFromString($name, $file);
-                continue;
-            }
-            
-            $songPath = $this->config->get('song_tmp_path') . "/" . $song['temporal_file_name'];
-            $trackNumber = sprintf("%02s", $song['trackNumber']);
-            $ext = pathinfo($songPath, PATHINFO_EXTENSION);
-            $name = "$trackNumber - {$song['name']}.$ext";
-            $zip->addFile($songPath, $name);
-        }
-        
-        $ext = pathinfo($this->tmpCoverImg, PATHINFO_EXTENSION);
-        $zip->addFile($this->tmpCoverImg, $this->config->get('cover_image_name') . '.'. $ext);
-        $zip->close();
-        
-        unlink($this->tmpCoverImg);
-        
-        echo "Subiendo zip\n";
-        
-        $response = $this->upload($filePath, $folder . '/' .$zipName);
-        unlink($filePath);
-        if(is_null($response)) {
-            throw new Exception('Error al intentar subir un archivo zip');
-        }
-        
-        return array('id' => $response['id'], 'name'=>$response['name'], 'url' => $response['mediaLink']);
-    }
-    
-    public function updateZip($album, $songs, $image = null) {
-        $this->downloadZip($album);
-        $zipPath = $this->config->get('zip_tmp_path') . '/' . $album['static_zip_file_name'];
-        
-        echo "Abriendo archivo zip\n";
-        $zip = new ZipArchive;
-        if ($zip->open($zipPath) !== TRUE) {
-            throw new Exception('Error al intentar abrir el archivo zip');
-        }
-        
-        echo "Actualizando canciones al archivo zip\n";
-        // add songs
-        foreach($songs['add'] as $song) {
-            $songPath = $this->config->get('song_tmp_path') . "/" . $song['temporal_file_name'];
-            $trackNumber = sprintf("%02s", $song['trackNumber']);
-            $ext = pathinfo($songPath, PATHINFO_EXTENSION);
-            $name = "$trackNumber - {$song['name']}.$ext";
-            $zip->addFile($songPath, $name);
-        }
-        // remove songs
-        foreach($songs['remove'] as $song) {
-            $toDelete = str_replace($album['staticDirectory'] . '/', '', $song);
-            $zip->deleteName($toDelete);
-        }
-        
-        // image
-        if(!is_null($image)) {
-            echo "Actualizando imagen del archivo zip\n";
-            $ext = pathinfo($this->tmpCoverImg, PATHINFO_EXTENSION);
-            $zip->addFile($this->tmpCoverImg, $this->config->get('cover_image_name') . '.'. $ext);
-        }
-        
-        $zip->close();
-        
-        echo "Subiendo zip\n";
-        $this->delete($album['static_zip_file_name']);
-        $response = $this->upload($zipPath, $album['staticDirectory'] . '/' .$album['name']);
-        unlink($zipPath);
-        if(is_null($response)) {
-            throw new Exception('Error al intentar subir un archivo zip');
-        }
-        
-        return array('id' => $response['id'], 'name'=>$response['name'], 'url' => $response['mediaLink']);
-    }
-    
 }
